% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{addTrail}
\alias{addTrail}
\alias{addTrail,CellTrailsSet-method}
\title{ADD trail}
\usage{
addTrail(ctset, from, to, name)
}
\arguments{
\item{ctset}{An object of class \code{\link[CellTrails]{CellTrailsSet}}}

\item{from}{Start node ID}

\item{to}{End node ID}

\item{name}{Name of trail}
}
\value{
An updated object of class \code{\link[CellTrails]{CellTrailsSet}}
}
\description{
Function to define a single trail on the trajectory.
}
\details{
A trajectory can be composed of multiple single trails (e.g., developmental
progression from a common start towards distinct terminal phenotypes).
Start and endpoints of trails can be identified using the plot function
(\code{plot(ctset, type = "trailblazing")}. Here, start (=from) and end (=to) IDs
of samples are starting with the character "B"
(for branching points) or "H" (for trail heads, i.e. terminal nodes).
\cr \cr
\emph{Diagnostic messages}
\cr \cr
An error is thrown if the trajectory has not been fitted yet. Please,
call \code{\link{fitTrajectory}} first. Further, an error is thrown if the
provided start or end ID is unknown.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Select trajectory
ctset <- selectTrajectory(ctset, component=1)

# Align samples to trajectory
ctset <- fitTrajectory(ctset)

# For illustration purposes a layout for each example dataset
# is contained in this package.
# The function trajectoryLayout allows to set a trajectory layout
# to the object; the parameter adjust='TRUE' adjusts the layout
# according to the computed pseudotime
trajectoryLayout(ctset, adjust=TRUE) <- trajLayouts$example

# Define trail
ctset <- addTrail(ctset, from="H1", to="H3", name="Tr1") # Define trail
}
\author{
Daniel C. Ellwanger
}
