% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor-methods.R
\docType{methods}
\name{as.ExpressionSet}
\alias{as.ExpressionSet}
\alias{as.ExpressionSet,CellTrailsSet-method}
\title{Coerce CellTrailsSet into ExpressionSet}
\usage{
as.ExpressionSet(object)
}
\arguments{
\item{object}{A \code{\link[CellTrails]{CellTrailsSet}} object}
}
\value{
An \code{\link[Biobase]{ExpressionSet}} object
}
\description{
Function to coerce a \code{\link[CellTrails]{CellTrailsSet}} object into
an \code{\link[Biobase]{ExpressionSet}} object.
}
\examples{
# Generate example data
dat <- exDat()
dat
ctset <- as.CellTrailsSet(dat)

# From CellTrailsSet to ExpressionSet
dat2 <- as.ExpressionSet(ctset)
dat2
}
\seealso{
ExpressionSet
}
