% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor-methods.R
\docType{methods}
\name{as.SingleCellExperiment}
\alias{as.SingleCellExperiment}
\alias{as.SingleCellExperiment,CellTrailsSet-method}
\title{Coerce CellTrailsSet into SingleCellExperiment}
\usage{
as.SingleCellExperiment(object)
}
\arguments{
\item{object}{A \code{CellTrailsSet} object}
}
\value{
A \code{SingleCellExperiment} object
}
\description{
Function to coerce a \code{CellTrailsSet} object into
a \code{SingleCellExperiment} object.
}
\details{
The resulting \code{SingleCellExperiment} object
contains additional \code{colData} information from the CellTrails
analysis (e.g., the determined states, the pseudotime
along each trail).
}
\examples{
# Generate example data
dat <- exDat()
dat
ctset <- as.CellTrailsSet(dat)

# From CellTrailsSet to ExpressionSet
dat2 <- as.ExpressionSet(ctset)
dat2
}
