% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellTrailsSpectrum}
\alias{CellTrailsSpectrum}
\alias{CellTrailsSpectrum-class}
\title{Class to determine the dimensions of the manifold}
\description{
Container for spectral embedding related information; not
to be initialized by user directly.
}
\section{Slots}{

\describe{
\item{\code{frac}}{Object of class \code{\link[base]{numeric}} containing the number of eigengaps used to perform linear fit}

\item{\code{n}}{Object of class \code{\link[base]{numeric}} containing the number of relevant dimensions}

\item{\code{cs}}{Object of class \code{\link[base]{numeric}} contaning the cumulative sum of eigengaps}

\item{\code{fit}}{Object of class \code{\link[base]{list}} containing the results of the linear fit}
}}

\author{
Daniel C. Ellwanger
}
