% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-methods.R
\name{.adjustLayoutByPtime}
\alias{.adjustLayoutByPtime}
\title{Adjusting the trajectory graph layout}
\usage{
.adjustLayoutByPtime(x, l)
}
\arguments{
\item{x}{A \code{SingleCellExperiment} object}

\item{layout}{Layout coordinates for each sample}
}
\value{
An adjusted layout
}
\description{
Adjusts the edge distances in the trajectory graph
such that edge distances correlates to edge weights;
pseudotime is stored in edge weights
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
