% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-methods.R
\name{.capitalize}
\alias{.capitalize}
\title{Capitalizes first character of string}
\usage{
.capitalize(x)
}
\arguments{
\item{x}{A string}
}
\value{
A string
}
\description{
Capitalizes first character of string and sets the rest to
lower case.
}
\details{
Example: "abC" becoms "Abc".
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
