% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-methods.R
\name{.color_hue}
\alias{.color_hue}
\title{Color palette}
\usage{
.color_hue(n)
}
\arguments{
\item{n}{Number of colors to be generated}
}
\value{
Colors codes
}
\description{
Generates equally spaced hues around the color wheel
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
