% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-methods.R
\name{.color_ramp}
\alias{.color_ramp}
\title{Colors for vector}
\usage{
.color_ramp(
  x,
  range = 3,
  colPal = NULL,
  min.val = 1e-10,
  min.val.col = NA,
  breaks = 25
)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Color codes
}
\description{
Generates colors for vector
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
