% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{.connect_ordi}
\alias{.connect_ordi}
\title{AUX Generate graph based on ordination}
\usage{
.connect_ordi(ordi)
}
\arguments{
\item{ordi}{Ordination}
}
\value{
An \code{igraph} object
}
\description{
Generates 2D ordination from orthogonal projection
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
