% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{.connect_ortho}
\alias{.connect_ortho}
\title{AUX: Connect orthogonal projected samples}
\usage{
.connect_ortho(Xorth)
}
\arguments{
\item{Xorth}{Result from the orthogonal projection}
}
\value{
An \code{igraph} object
}
\description{
Connects samples based on their projected position on the trajectory.
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
