% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{.connectStates_def}
\alias{.connectStates_def}
\title{DEF: Connect states}
\usage{
.connectStates_def(dmat, cl, l = 10, sigma = 1)
}
\arguments{
\item{dmat}{Distance matrix}

\item{cl}{States}
}
\description{
For details see \code{connectStates}
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
