% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp-methods.R
\name{.contrastExprTrail_def}
\alias{.contrastExprTrail_def}
\title{DEF: Differential expression between trails}
\usage{
.contrastExprTrail_def(
  ptime_1,
  ptime_2,
  feature_expr,
  sts,
  trail_names,
  n = 250,
  score,
  dynamicFit_k = 5,
  df = 10
)
}
\arguments{
\item{ptime_1}{Pseudotime trail 1}

\item{ptime_2}{Pseudotime trail 2}

\item{feature_expr}{Expression vector}

\item{sts}{Samples vector}

\item{n}{Number of fitted values}

\item{score}{Score type}

\item{dynamicFit_k}{Freedom param for dynamic fit}

\item{df}{Freedom param for splines fit}
}
\description{
DEF: Differential expression between trails
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
