% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{.deleteMedianCentres}
\alias{.deleteMedianCentres}
\title{AUX Remove median centres}
\usage{
.deleteMedianCentres(g)
}
\arguments{
\item{g}{An object of class \code{igraph}}
}
\value{
An unpdated object of class \code{igraph}
}
\description{
Deletes median centres from trajectory graph
if graph has not been simplified/has been expanded.
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
