% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp-methods.R
\name{.diffExprState_def}
\alias{.diffExprState_def}
\title{DEF: Differential expression between states}
\usage{
.diffExprState_def(
  x,
  state1,
  state2,
  feature_name,
  alternative = "two.sided",
  lod = NULL
)
}
\description{
DEF: Differential expression between states
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
