% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred-methods.R
\name{.embedSamples_def}
\alias{.embedSamples_def}
\title{DEF: Spectral embedding of samples}
\usage{
.embedSamples_def(x, nbins = 10, sigma = 0.75, design = NULL)
}
\arguments{
\item{x}{A numerical matrix}

\item{nbins}{Cubic B-spline discretization is used to compute fuzzy mutual
information between pairs of samples; \code{nbins} defines the number
of intervals used for discretization of expression data. (default: 10)}

\item{sigma}{The mutual information matrix is transformed to an
adjacency matrix using a heat kernel; \code{sigma} defines the radius of
heat kernel (in quantiles; default: \code{sigma} = 0.75 which is the third
quantile of the mutual information matrix).}

\item{design}{A numeric matrix describing the factors that should be blocked}
}
\description{
For details see \code{embedSamples}
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
