% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.exprs}
\alias{.exprs}
\alias{.exprs,SingleCellExperiment-method}
\title{GET expression matrix}
\usage{
.exprs(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A \code{numeric} matrix
}
\description{
Retrieve numeric matrix of expression values for processing in
CellTrails. This wrapper function ensures that all functions in the
package receive the proper assay from the
\code{SingleCellExperiment} object.
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
