% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-methods.R
\name{.featureNameExists}
\alias{.featureNameExists}
\title{Checks if feature exists}
\usage{
.featureNameExists(x, feature_name)
}
\arguments{
\item{x}{A \code{SingleCellExperiment} object}

\item{feature_name}{Name of feature}
}
\value{
logical value
}
\description{
Checks if feature exists
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
