% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-methods.R
\name{.filterTrajFeaturesByCOV_def}
\alias{.filterTrajFeaturesByCOV_def}
\title{DEF: Filter features by coefficient of variation}
\usage{
.filterTrajFeaturesByCOV_def(y, threshold, design = NULL, show_plot = TRUE)
}
\arguments{
\item{y}{An expression vector}

\item{threshold}{numeric cutoff value}

\item{design}{Model matrix}

\item{show_plot}{Show plot?}
}
\description{
For details see \code{filterFeaturesByCOV}
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
