% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-methods.R
\name{.filterTrajFeaturesByDL_def}
\alias{.filterTrajFeaturesByDL_def}
\title{DEF: Filter feaures by Detection Level}
\usage{
.filterTrajFeaturesByDL_def(y, threshold, show_plot = TRUE)
}
\arguments{
\item{y}{An expression matrix}

\item{threshold}{numeric cutoff value}

\item{show_plot}{Show plot?}
}
\description{
For details see \code{filterFeaturesByPOD}
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
