% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-methods.R
\name{.filterTrajFeaturesByFF_def}
\alias{.filterTrajFeaturesByFF_def}
\title{DEF: Filter features by index of dispersion / fano factor}
\usage{
.filterTrajFeaturesByFF_def(
  y,
  z,
  min_expr = 0,
  design = NULL,
  show_plot = TRUE
)
}
\arguments{
\item{y}{An expression vector}

\item{z}{A cutoff z-score}

\item{min_expr}{Minimum expression level}

\item{design}{Model matrix}

\item{show_plot}{Show plot?}
}
\description{
For details see \code{filterFeaturesByFF}
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
