% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred-methods.R
\name{.findSpectrum_def}
\alias{.findSpectrum_def}
\title{DEF: Determine number of informative latent dimensions}
\usage{
.findSpectrum_def(D, frac = 100)
}
\description{
For details see \code{findSpectrum}
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
