% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-methods.R
\name{.fitDynamic_def}
\alias{.fitDynamic_def}
\title{Fitting dynamics}
\usage{
.fitDynamic_def(x, y, z, k = 5, n.out = NULL, x.out = NULL)
}
\arguments{
\item{x}{Pseudotime values}

\item{y}{Expression values}

\item{z}{Sample states}

\item{k}{Dimensions of the bases used to represent the smooth term}

\item{n.out}{Number of predicted expression values within pseudotime range}

\item{x.out}{Predicted expression values for given set of pseudotime values}
}
\value{
A list containing the following components:
\item{\code{x}}{Pseudotime}
\item{\code{y}}{Fitted values}
\item{\code{mod}}{GAM}
}
\description{
Fits expression as a function of pseudotime using generalized
additive models
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
