% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{.fitTrajectory_def}
\alias{.fitTrajectory_def}
\title{DEF: Trajectory fitting}
\usage{
.fitTrajectory_def(cl = cl, g = g, X = X, snames)
}
\arguments{
\item{cl}{Trajectory states vector}

\item{g}{Trajectory graph; \code{igraph} object}

\item{X}{Lower-dimensional ordination of trajectory samples}

\item{snames}{Sample names}
}
\description{
For details see \code{fitTrajectory}
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
