% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-methods.R
\name{.fr_layout}
\alias{.fr_layout}
\title{Computes state trajectory graph layout}
\usage{
.fr_layout(g)
}
\arguments{
\item{g}{An igraph graph object}
}
\value{
numerical matrix with the layout coordinates
}
\description{
Uses the Fruchterman-Reingold layout algorithm
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
