% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{.generate_ordination}
\alias{.generate_ordination}
\title{AUX: Orthogonal projection ordination}
\usage{
.generate_ordination(g, error, factor = 7, rev = FALSE, only.ordi = FALSE)
}
\arguments{
\item{g}{Graph from orthogonal projection, \code{igraph} object}

\item{error}{Error from orthogonal projection; \code{numeric} vector}

\item{only.ordi}{Compute only ordination?}
}
\value{
A list containing the following components:

\item{\code{ordi}}{Ordination}

\item{\code{ordi.jitter}}{Jittered ordination}

\item{\code{backbone}}{Backbone of trajectory}
}
\description{
Generates 2D ordination from orthogonal projection
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
