% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-methods.R
\name{.ihs}
\alias{.ihs}
\title{Reverse inverse hyperbolic sine}
\usage{
.ihs(x)
}
\arguments{
\item{x}{A numeric value, vector or matrix}
}
\value{
Transformed values
}
\description{
Computes inverse hyperbolic sine
}
\details{
In contrast to sqrt and log, ihs is defined for negative numbers.
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
