% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-methods.R
\name{.linear_fit}
\alias{.linear_fit}
\title{Linear fit}
\usage{
.linear_fit(x.in, y.in, x.out = NULL)
}
\arguments{
\item{x.in}{Predictor variable, numeric vector}

\item{y.in}{Response variable, numeric vector}

\item{x.out}{Predictor variable for linear function, numeric vector}
}
\value{
A list containing the following components:

\item{\code{p.value}}{Anova P-value}
\item{\code{r2}}{Adjusted R-squared}
\item{\code{x.out}}{Predictor variable for linear function, numeric vector}
\item{\code{y.out}}{Response from linear function}
}
\description{
Perform simple linear fit
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
