% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-methods.R
\name{.nn_impute}
\alias{.nn_impute}
\title{Nearest neighbor imputation}
\usage{
.nn_impute(y, D)
}
\arguments{
\item{y}{Values; missing values need to be set to NA}

\item{D}{distance matrix between samples}
}
\value{
Imputed values
}
\description{
Nearest neighbor imputation
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
