% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred-methods.R
\name{.pca_def}
\alias{.pca_def}
\title{DEF: PCA}
\usage{
.pca_def(M, do_scaling = TRUE, design = NULL)
}
\arguments{
\item{M}{expression matrix}

\item{do_scaling}{use covariance/correlation matrix}

\item{design}{Block factors}
}
\description{
For details see \code{pca}
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
