% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.pheno}
\alias{.pheno}
\alias{.pheno,SingleCellExperiment-method}
\title{GET phenotype values}
\usage{
.pheno(object, name)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}

\item{name}{Name of phenotype}
}
\value{
A vector of any type
}
\description{
Returns phenotype values from
\code{SingleCellExperiment} object
}
\details{
Wrapper for colDat(object)[, name] which also accesses
internal metadata (e.g., landmarks).
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
