% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-methods.R
\name{.phenoNameExists}
\alias{.phenoNameExists}
\title{Checks if phenotype exists}
\usage{
.phenoNameExists(x, pheno_name)
}
\arguments{
\item{x}{A \code{SingleCellExperiment} object}

\item{pheno_name}{Name of phenotype}
}
\value{
logical value
}
\description{
Checks if phenotype exists
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
