% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{.plot_trajectoryFit}
\alias{.plot_trajectoryFit}
\title{DEF: Visualizing the trajectory fit}
\usage{
.plot_trajectoryFit(x, g, sts, factor = 7, rev = FALSE)
}
\arguments{
\item{x}{The fitting residuals}

\item{g}{The trajectory graph}

\item{sts}{The states along the trajectory}

\item{factor}{The jitter intensity correlates to the projection error}

\item{rev}{Should the trajectory shown in reverse order?}
}
\value{
A \code{ggplot} object
}
\description{
Shows the residuals along the trajectory backbone
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
