% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{.plotDynamic}
\alias{.plotDynamic}
\title{DEF: Visualize expression dynamic}
\usage{
.plotDynamic(x, Y, feature_name, trail_name, weights, k = 5)
}
\arguments{
\item{x}{Pseudotime}

\item{Y}{Expression matrix (samples in columns)}

\item{weights}{Weight (states)}

\item{k}{Dimensions of bases used to represent the smooth term}
}
\description{
Shows feature expression as a function of pseudotime
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
