% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{.plotManifold_def}
\alias{.plotManifold_def}
\title{DEF: Visualizing the manifold}
\usage{
.plotManifold_def(
  X,
  y,
  name,
  g = NULL,
  weights = NULL,
  axis_label = "",
  colors = c("pretty", "bw"),
  type = c("raw", "surface.fit", "surface.se"),
  samples_only = FALSE,
  setND = FALSE,
  alpha = 0.1
)
}
\arguments{
\item{X}{Ordination of data points}

\item{y}{Values to visualize}

\item{name}{Label for the figure legend}

\item{g}{The trajectory graph}

\item{weights}{States along the trajectory}

\item{axis_label}{Label prefix for each axis}

\item{colors}{Either the color ramp is colorized or black/white}

\item{type}{Type of visualization}

\item{samples_only}{Does not show the whole surface, but colorizes
only the samples}

\item{setND}{Indicates if a '0' value should be set to 'ND' in the
figure legend}

\item{alpha}{Color param}
}
\value{
A \code{ggplot} object
}
\description{
Plots approximation of sample embedding in latent space in two dimensions
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
