% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{.plotSpectrum_def}
\alias{.plotSpectrum_def}
\title{DEF: Visualizing the spectrum definition}
\usage{
.plotSpectrum_def(x)
}
\arguments{
\item{x}{A \code{list}}
}
\value{
A \code{ggplot} object
}
\description{
Plots Scree plot with eigengaps
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
