% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{.plotStateExpression_def}
\alias{.plotStateExpression_def}
\title{DEF: Violine plots of genes}
\usage{
.plotStateExpression_def(x, sts, label)
}
\arguments{
\item{x}{A expression vector}

\item{sts}{A states assignment vector}

\item{label}{Label for legend}
}
\value{
A \code{ggplot} object
}
\description{
Plot gene expression per state
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
