% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{.plotStateSize_def}
\alias{.plotStateSize_def}
\title{DEF: Visualizing states sizes}
\usage{
.plotStateSize_def(x)
}
\arguments{
\item{x}{States}
}
\value{
A \code{ggplot} object
}
\description{
Plots barplot of number of samples per state
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
