% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{.plotStateTrajectory_def}
\alias{.plotStateTrajectory_def}
\title{DEF: Visualizing the trajectory graph}
\usage{
.plotStateTrajectory_def(
  X,
  g,
  y,
  all_sts,
  name,
  setND = FALSE,
  point_size = 3,
  label_offset = 2
)
}
\arguments{
\item{X}{Layout}

\item{g}{State trajectory graph (igraph object)}

\item{y}{Values; numeric or factor vector}

\item{all_sts}{All state assignments}

\item{setND}{If values of '0' are getting set to 'ND' in the
figure legend}

\item{point_size}{Point size paramenter}

\item{label_offset}{Label offset parameter}

\item{label}{Label for plot legend}
}
\value{
A \code{ggplot} object
}
\description{
Visualizes the trajectory spanning all states of a component
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
