% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{.plotTrail_def}
\alias{.plotTrail_def}
\title{DEF: Visualize trail on map}
\usage{
.plotTrail_def(X, g, ptime, name)
}
\arguments{
\item{X}{The trajectory layout}

\item{g}{The trajectory graph}

\item{ptime}{The trail pseudotime}

\item{name}{Name of trail}
}
\description{
Highlights a trail on map
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
