% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{.plotTrailblazing_def}
\alias{.plotTrailblazing_def}
\title{DEF: Visualizing trailblazing}
\usage{
.plotTrailblazing_def(X, g, ltype, lid)
}
\arguments{
\item{X}{The trajectory layout}

\item{g}{The trajectory graph}

\item{ltype}{Vector with landmark types}

\item{lid}{Vector with landmark ids}
}
\value{
A \code{ggplot} object
}
\description{
Shows trajectory graph and highlights landmarks
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
