% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-methods.R
\name{.prettyColorRamp}
\alias{.prettyColorRamp}
\title{Pretty color ramp}
\usage{
.prettyColorRamp(n, grayStart = TRUE)
}
\arguments{
\item{n}{Number of colors}
}
\value{
Color codes
}
\description{
Pretty color ramp
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
