% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-methods.R
\name{.prettyString}
\alias{.prettyString}
\title{Pretty string from array}
\usage{
.prettyString(x, mmax = 4)
}
\arguments{
\item{x}{A character vector}
}
\value{
String
}
\description{
Generates short representation of long
character vectors
}
\details{
Returns pretty print of character vector
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
