% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{.project_ortho}
\alias{.project_ortho}
\title{AUX: Orthogonal projection}
\usage{
.project_ortho(cl, g, X)
}
\arguments{
\item{cl}{Trajectory states; factor vector}

\item{g}{Trajectory tree; igraph object}

\item{X}{Lower dimensional ordination of trajectory samples}
}
\value{
A \code{list} containing the following components:
\item{\code{Y}}{Projection}
\item{\code{X.c}}{Mediancentres}
\item{\code{lambda}}{Projection lambda (position: inside/outside of edge)}
\item{\code{adjmat}}{Adjacency matrix with sample count per edge}
\item{\code{error}}{Projection error per sample}
\item{\code{edgeId}}{ID of each edge}
\item{\code{edgeId2Edge}}{Mapping of edge to edgeId}
\item{\code{edge}}{For each sample its assigned edge}
}
\description{
Orthogonally projects samples onto trajectory
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
