% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-methods.R
\name{.rbf}
\alias{.rbf}
\title{Radial basis function}
\usage{
.rbf(d, radius)
}
\arguments{
\item{x}{A numeric value, vector or matrix}

\item{sigma}{Radius of the kernel}
}
\value{
Transformed values
}
\description{
Computes radial basis function
}
\details{
Also known as Heat kernel or Gaussian kernel
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
