% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-methods.R
\name{.rescale}
\alias{.rescale}
\title{Rescale vector}
\usage{
.rescale(x, ymin, ymax)
}
\arguments{
\item{x}{X-values}

\item{ymin}{New min}

\item{ymax}{New max}
}
\value{
Rescaled vector
}
\description{
Rescales vector to [ymin, ymax]
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
