% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-methods.R
\name{.sampleNameExists}
\alias{.sampleNameExists}
\title{Checks if sample exists}
\usage{
.sampleNameExists(x, sample_name)
}
\arguments{
\item{x}{A \code{SingleCellExperiment} object}

\item{sample_name}{Name of sample}
}
\value{
logical value
}
\description{
Checks if sample exists
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
