% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.spanForest<-}
\alias{.spanForest<-}
\alias{.spanForest<-,SingleCellExperiment-method}
\title{SET state trajectory graph}
\usage{
.spanForest(object) <- value
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}

\item{value}{A \code{list} object with an \code{igraph} object
per component of the spanning forest}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Sets graph object spanning all states (spanning forest)
to \code{SingleCellExperiment} object
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
