% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.spanForest}
\alias{.spanForest}
\alias{.spanForest,SingleCellExperiment-method}
\title{GET state trajectory graph}
\usage{
.spanForest(object)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}
}
\value{
A \code{list} object with an \code{igraph} object
per component of the spanning forest
}
\description{
Returns graph object spanning all states (spanning forest)
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
