% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-methods.R
\name{.spatmed}
\alias{.spatmed}
\title{Spatial median}
\usage{
.spatmed(X, maxiter = 500, eps = 1e-06)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A numeric vector
}
\description{
Computes mediancentres
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
