% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.stateTrajLayout}
\alias{.stateTrajLayout}
\alias{.stateTrajLayout,SingleCellExperiment-method}
\title{GET state trajectory layout}
\usage{
.stateTrajLayout(object, component)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}
}
\value{
A \code{numeric} matrix
}
\description{
Gets layout of state trajectory from
\code{SingleCellExperiment} object
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
