% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-methods.R
\name{.trailNameExists}
\alias{.trailNameExists}
\title{Checks if trail exists}
\usage{
.trailNameExists(x, trail_name)
}
\arguments{
\item{x}{A \code{SingleCellExperiment} object}

\item{trail_name}{Name of trail}
}
\value{
logical value
}
\description{
Checks if trail exists
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
