% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.trajGraph<-}
\alias{.trajGraph<-}
\alias{.trajGraph<-,SingleCellExperiment-method}
\title{SET trajectory graph}
\usage{
.trajGraph(object) <- value
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}

\item{value}{A \code{igraph} object}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Stores trajectory graph in a \code{SingleCellExperiment} object
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
