% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.trajGraph}
\alias{.trajGraph}
\alias{.trajGraph,SingleCellExperiment-method}
\title{GET trajectory graph}
\usage{
.trajGraph(object)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}
}
\value{
A \code{igraph} object
}
\description{
Returns trajectory graph from a \code{SingleCellExperiment} object
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
