% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.trajLandmark<-}
\alias{.trajLandmark<-}
\alias{.trajLandmark<-,SingleCellExperiment-method}
\title{SET trajectory landmark annotation}
\usage{
.trajLandmark(object, type = c("type", "id", "shape")) <- value
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}

\item{type}{A character in {"type", "id", "shape"}}

\item{value}{A vector of any type}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Stores information on trajectory landmarks
in a \code{SingleCellExperiment} object
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
