% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.trajLandmark}
\alias{.trajLandmark}
\alias{.trajLandmark,SingleCellExperiment-method}
\title{GET trajectory landmark annotation}
\usage{
.trajLandmark(object, type = c("type", "id", "shape"))
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}

\item{type}{A character in {"type", "id", "shape"}}
}
\value{
A vector of any type
}
\description{
Returns trajectory landmark information from
\code{SingleCellExperiment} object
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
