% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.trajResiduals<-}
\alias{.trajResiduals<-}
\alias{.trajResiduals<-,SingleCellExperiment-method}
\title{SET trajectory fitting residuals}
\usage{
.trajResiduals(object) <- value
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}

\item{value}{A \code{numeric} vector}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Stores trajectory fitting residuals in \code{SingleCellExperiment} object
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
