% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.useFeature<-}
\alias{.useFeature<-}
\alias{.useFeature<-,SingleCellExperiment-method}
\title{SET trajectory features indicator}
\usage{
.useFeature(object) <- value
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}

\item{value}{A logical vector}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Sets indicator if feature should be used for trajectory
reconstruction.
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
