% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.useFeature}
\alias{.useFeature}
\alias{.useFeature,SingleCellExperiment-method}
\title{GET trajectory features indicator}
\usage{
.useFeature(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A \code{logical} vector
}
\description{
Indicates if feature should be used for trajectory
reconstruction. Spike-in controls are removed.
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
