% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{.useSample}
\alias{.useSample}
\alias{.useSample,SingleCellExperiment-method}
\title{GET trajectory samples indicator}
\usage{
.useSample(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A \code{logical} vector
}
\description{
Indicates if sample was used for trajectory reconstruction.
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
