% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-methods.R
\name{.validatePlotParams}
\alias{.validatePlotParams}
\title{Validates parameter settings for plot methods}
\usage{
.validatePlotParams(x, color_by, name)
}
\arguments{
\item{x}{A \code{SingleCellExperiment} object}

\item{color_by}{Color_by parameter}

\item{name}{Name parameter}
}
\value{
A \code{list} with the parameters and values
}
\description{
Validates parameter settings for plot methods
}
\details{
Parameter \code{color_by} needs to be in
{'featureName', 'phenoName'}. Parameter \code{name} needs
to be existent in rownames(object) or in
{colnames(colData(object)), state, landmark}.
If parameters were validated true, this function automatically
extracts and returns the values. Character vectors get converted
to factors.
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
