% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\name{.write_ygraphml_def}
\alias{.write_ygraphml_def}
\title{DEF: Export trajectory graph}
\usage{
.write_ygraphml_def(g, X, file, col_values, lbl_values, shapes)
}
\arguments{
\item{g}{Trajectory graph}

\item{X}{Lower-dimensional ordination of trajectory samples}

\item{col_values}{Color values}

\item{lbl_values}{Label values}

\item{shapes}{Trajectory landmark shapes}

\item{tlayout}{Trajectory layout (if existent)}
}
\description{
For details see \code{write.ygraphml}
}
\author{
Daniel C. Ellwanger
}
\keyword{internal}
