% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{eigenvalues<-}
\alias{eigenvalues<-}
\alias{eigenvalues<-,CellTrailsSet-method}
\title{SET eigenvalues}
\usage{
eigenvalues(object) <- value
}
\arguments{
\item{object}{An object of class \code{CellTrailsSet}}

\item{value}{A numeric vector}
}
\value{
An updated object of class \code{CellTrailsSet}
}
\description{
Sets eigenvalues to a \code{CellTrailsSet} object.
}
\details{
Eigenvalues represent the information content for each
eigenvector derived by spectral embedding of samples. Vector
needs to be numeric and its length needs to correspond to the number
of eigenvectors. Usually, this function is not directly accessed
by the user.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Perform PCA and set results to object
pca_result <- pca(ctset)
latentSpace(ctset) <- pca_result$princomp
eigenvalues(ctset) <- pca_result$variance
}
\author{
Daniel C. Ellwanger
}
