% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{eigenvalues}
\alias{eigenvalues}
\alias{eigenvalues,CellTrailsSet-method}
\title{GET eigenvalues}
\usage{
eigenvalues(object)
}
\arguments{
\item{object}{An object of class \code{CellTrailsSet}}
}
\value{
An object of class \code{numeric}
}
\description{
Retrieve computed eigenvalues from a \code{CellTrailsSet} object.
}
\details{
Eigenvalues represent the information content for each eigenvector derived by
spectral embedding of samples. This function retrieves the derived eigenvalues from a
\code{CellTrailsSet} object. The corresponding vector is numeric.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Get eigenvalues
eigenvalues(ctset)
}
\author{
Daniel C. Ellwanger
}
