% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-methods.R
\docType{methods}
\name{exDat}
\alias{exDat}
\title{Simulation of example data for trajectory reconstruction}
\usage{
exDat()
}
\value{
A \code{SingleCellExperiment} object
}
\description{
This method serves to generate the example data used to
demonstrate the usage of individual functions of this package
}
\details{
RNA-Seq counts are generated using the Negative Binomial
Distribution. Distribution parameters for each feature are sampled from a
Gamma distribution. The resulting expression matrix is then log2-scaled.
The assay data consist of 25 features and 100 samples.
}
\examples{
# Generate example data
exDat()
}
\seealso{
\code{SingleCellExperiment} \code{simulate_exprs}
}
\author{
Daniel C. Ellwanger
}
