% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{featureNames}
\alias{featureNames}
\alias{featureNames,SingleCellExperiment-method}
\title{GET feature names}
\usage{
featureNames(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A \code{character} vector
}
\description{
Retrieve feature names from a \code{SingleCellExperiment} object
}
\details{
Wrapper for \code{rownames(object)}
}
\examples{
# Example data
data(exSCE)

featureNames(exSCE)
}
\seealso{
\code{SingleCellExperiment}
}
\author{
Daniel C. Ellwanger
}
