% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{filterFeaturesByCOV}
\alias{filterFeaturesByCOV}
\alias{filterFeaturesByCOV,CellTrailsSet-method}
\title{Filter features by Coefficient of Variation (COV)}
\usage{
filterFeaturesByCOV(ctset, threshold, design = NULL)
}
\arguments{
\item{ctset}{An \code{CellTrailsSet} object}

\item{threshold}{Minimum coefficient of variation;
numeric value between 0 and 1}

\item{design}{A numeric matrix describing the factors that should be blocked}
}
\value{
An updated \code{CellTrailsSet} object
}
\description{
Filters trajectory features that have a minimal coefficient of variation.
}
\details{
For each trajectory feature \emph{x} listed in the CellTrailsSet
object the coefficient of variation is computed by
\eqn{CoV(x) = sd(x) / mean(x)}. Features with a CoV(x) greater
than \code{threshold} remain labeled as trajectory feature in the
\code{CellTrailsSet} object, otherwise they are not considered
for dimensionality reduction, clustering and trajectory reconstruction.
\cr \cr
To account for systematic bias in the expression data
(e.g., cell cycle effects), a design matrix can be provided for the
learning process. It should list the factors that should be blocked and
their values per sample. It is suggested to construct a design
matrix with \code{model.matrix}.
}
\examples{
# Simulate example data
dat <- simulate_exprs(n_features=15000, n_samples=100)

# Create container
ctset <- as.CellTrailsSet(dat)

# Filter
ctset <- filterFeaturesByDL(ctset, threshold=2)
ctset <- filterFeaturesByCOV(ctset, threshold=0.5)

length(trajectoryFeatures(ctset))
}
\seealso{
\code{\link[CellTrails]{CellTrailsSet}}
\code{\link[CellTrails]{trajectoryFeatures}}
\code{\link[stats]{model.matrix}}
}
\author{
Daniel C. Ellwanger
}
