% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{filterFeaturesByDL}
\alias{filterFeaturesByDL}
\alias{filterFeaturesByDL,CellTrailsSet-method}
\title{Filter features by Detection Level (DL)}
\usage{
filterFeaturesByDL(ctset, threshold)
}
\arguments{
\item{ctset}{An \code{\link[CellTrails]{CellTrailsSet}} object}

\item{threshold}{Minimum number of samples; if value < 1 it is interpreted as
fraction, otherwise as absolute sample count}
}
\value{
An updated \code{\link[CellTrails]{CellTrailsSet}} object
}
\description{
Filters trajectory features that are detected in a minimum number of samples.
}
\details{
The detection level denotes the fraction of samples in which a feature was detected.
For each trajectory feature listed in the
CellTrailsSet object the relative number of samples having a feature expression
value greater than 0 is counted. Features that are expressed in a fraction of all
samples greater than \code{threshold} remain labeled as trajectory feature in
the \code{CellTrailsSet} object, otherwise
they are not considered for dimensionality reduction, clustering and trajectory reconstruction.
If the parameter \code{threshold} fullfills
\code{threshold} \eqn{>= 1} it becomes converted to a relative fraction of the
total sample count.
}
\examples{
# Simulate example data
dat <- simulate_exprs(n_features=15000, n_samples=100)

# Create container
ctset <- as.CellTrailsSet(dat)

# Filter features
ctset <- filterFeaturesByDL(ctset, threshold=2)

length(trajectoryFeatures(ctset))
}
\seealso{
\code{\link{trajectoryFeatures}}
}
\author{
Daniel C. Ellwanger
}
