% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{gga_e15_utricle}
\alias{gga_e15_utricle}
\title{Example single-cell RT-qPCR data}
\format{An object of class \code{\link[Biobase]{ExpressionSet}}}
\usage{
data("gga_e15_utricle")
}
\description{
This dataset contains transcript expression profiles of 183 genes
expressed during sensory hair cell bundle maturation and function,
which were quantified in the chicken utricle sensory epithelium at
embryonic day 15 using multiplex RT-qPCR. Experimental metadata was
generated during tissue preparation (cell origin) and cell sorting
(uptake of FM1-43 dye indicating cell maturity). This data set is the
foundation used for the development of CellTrails.
}
\references{
Ellwanger DC, Scheibinger M, Dumont RA,
Barr-Gillespie PG, and Heller S. "Transcriptional dynamics of
hair-bundle morphogenesis revealed with CellTrails".
Cell Reports Date;Issue. doi:
}
\keyword{datasets}
