% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{landmarks}
\alias{landmarks}
\alias{landmarks,SingleCellExperiment-method}
\title{GET landmarks}
\usage{
landmarks(object)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}
}
\value{
A character vector with sample names
}
\description{
Gets landmarks from a \code{SingleCellExperiment} object.
}
\details{
Trail branches (B) and heads (H) are automatically assigned;
landmarks can also be defined on the trajectory by the user (U).
Landmarks can be used to extract single trails from a trajectory.
}
\examples{
# Example data
data(exSCE)

# Get landmarks
landmarks(exSCE)[seq_len(5)]
}
\seealso{
\code{userLandmarks}
}
\author{
Daniel C. Ellwanger
}
