% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{latentSpace<-}
\alias{latentSpace<-}
\alias{latentSpace<-,CellTrailsSet-method}
\title{SET latent space}
\usage{
latentSpace(object) <- value
}
\arguments{
\item{object}{An object of class \code{CellTrailsSet}}

\item{value}{A numeric matrix}
}
\value{
An updated object of class \code{CellTrailsSet}
}
\description{
Set computed latent space to a \code{CellTrailsSet} object.
}
\details{
Rows need to be samples and columns to be \emph{d} components
(spanning the lower-dimensional latent space). Usually, this function is not
directly accessed by the user.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Perform PCA and set components to object
pca_result <- pca(ctset)
latentSpace(ctset) <- pca_result$princomp
}
\author{
Daniel C. Ellwanger
}
