% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{manifold2D<-}
\alias{manifold2D<-}
\alias{manifold2D<-,SingleCellExperiment-method}
\title{SET 2D manifold representation}
\usage{
manifold2D(object) <- value
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}

\item{value}{A \code{numeric} matrix with one column per dimension}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Stores 2D manifold representation in \code{SingleCellExperiment} object
}
\examples{
# Example data
data(exSCE)

gp <- plotManifold(exSCE, color_by="featureName", name="feature_10",
                   recalculate=TRUE)
manifold2D(exSCE) <- gp
}
\author{
Daniel C. Ellwanger
}
