% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{manifold2D}
\alias{manifold2D}
\alias{manifold2D,SingleCellExperiment-method}
\title{GET 2D manifold representation}
\usage{
manifold2D(object)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}
}
\value{
A \code{numeric} matrix
}
\description{
Returns 2D manifold representation of latent space from
\code{SingleCellExperiment} object
}
\examples{
# Example data
data(exSCE)

manifold2D(exSCE)[seq_len(5), ]
}
\author{
Daniel C. Ellwanger
}
