% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{phenoNames}
\alias{phenoNames}
\alias{phenoNames,SingleCellExperiment-method}
\title{GET phenotype names}
\usage{
phenoNames(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A \code{character} vector
}
\description{
Retrieve phenotype names from a \code{SingleCellExperiment} object
}
\details{
Wrapper for \code{colnames(colData(object))}
}
\examples{
# Example data
data(exSCE)

phenoNames(exSCE)
}
\seealso{
\code{SingleCellExperiment}
}
\author{
Daniel C. Ellwanger
}
