% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plot.CellTrailsSpectrum}
\alias{plot.CellTrailsSpectrum}
\title{Visualize a CellTrailsSpectrum object}
\usage{
\method{plot}{CellTrailsSpectrum}(x, ...)
}
\arguments{
\item{x}{A \code{CellTrailsSpectrum} object}

\item{...}{Additional arguments; not normally accessed by user directly}
}
\value{
A \code{ggplot} object
}
\description{
Method illustrates the automatic determination of informative latent
dimensions.
}
\details{
Line segement plot showing the lagged differences between
ordered top \emph{frac} eigenvalues (eigengaps) and the linear fit.
Values above the linear fit are highlighted and corresponding eigenvalues
were considered to be part of the informative spectrum.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

#Plot
ggp <- plot(spectr)
ggp
}
\seealso{
\code{\link[ggplot2]{ggplot2-package}}
}
\author{
Daniel C. Ellwanger
}
