% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plotStateTrajectory}
\alias{plotStateTrajectory}
\alias{plotStateTrajectory,SingleCellExperiment-method}
\title{Visualize state trajectory graph}
\usage{
plotStateTrajectory(sce, color_by = c("phenoName", "featureName"), name,
  seed = 1101, component = NULL, point_size = 3, label_offset = 2)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}

\item{color_by}{Indicates if nodes are colorized by a feature expression
('featureName') or phenotype label ('phenoName')}

\item{name}{A character string specifying the featureName or phenoName}

\item{seed}{Seed for layout computation (default: 1101)}

\item{component}{Component of trajectory graph that should
be shown (integer value)}

\item{point_size}{Adjusts the point size of the data points shown}

\item{label_offset}{Adjusts the offset of the data point labels}
}
\value{
A \code{ggplot} object
}
\description{
Method visualizes the state-to-state relations delineating the
trajectory backbone.
}
\details{
Shows a single tree component of the computed trajectory graph.
Each point in this plot represents a state and can be colorized
according to feature expression (mean expression per state) or experimental
metadata (arithmetic mean or percentage distribution of categorial values).
The component is defined by parameter \code{component}. If the trajectory
graph contains only a single component, then this parameter can be left
undefined. The points' coloration can be defined via the attributes
\code{feature_name} or \code{pheno_type}. Missing sample lables are
recovered using nearest neighbor learning.
}
\examples{
# Generate example data
sce <- exDat()

# Reduce dimensionality
res <- embedSamples(sce)
d <- findSpectrum(res$eigenvalues, frac=30)
latentSpace(sce) <- res$components[, d]

# Find and connect states
states(sce) <- findStates(sce, max_pval=1e-3, min_feat=4)
sce <- connectStates(sce, l=15)

plotStateTrajectory(sce, color_by="phenoName", name="age", component=1)
plotStateTrajectory(sce, color_by="featureName", name="feature_1",
                    component=1)
}
\seealso{
\code{connectStates}
}
\author{
Daniel C. Ellwanger
}
