% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plotTrajectoryFit}
\alias{plotTrajectoryFit}
\alias{plotTrajectoryFit,SingleCellExperiment-method}
\title{Visualize trajectory fit residuals}
\usage{
plotTrajectoryFit(sce)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}
}
\value{
A \code{ggplot} object
}
\description{
Method visualizes the fitting residuals along the
trajectory backbone.
}
\details{
Shows the trajectory backbone (longest shortest path
between two samples) and the fitting deviations of each sample
indicated by the perpendicular jitter. Data points are colorized
by state.
}
\examples{
# Example data
data(exSCE)

plotTrajectoryFit(exSCE)
}
\seealso{
\code{fitTrajectory} \code{trajResiduals}
}
\author{
Daniel C. Ellwanger
}
