% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{reduceDimensions}
\alias{reduceDimensions}
\alias{reduceDimensions,CellTrailsSet-method}
\title{Dimensionality reduction}
\usage{
reduceDimensions(ctset, ctspec)
}
\arguments{
\item{ctset}{A \code{CellTrailsSet} object}

\item{ctspec}{A \code{CellTrailsSpectrum} object}
}
\value{
An updated object of class \code{CellTrailsSet}
}
\description{
Truncates eigenbasis based on the eigengap criterion.
}
\details{
The number of informative dimensions was determined
using the function \code{findSpectrum}. The resulting
object is of class \code{CellTrailsSpectrum} and is used
in this method to reduce the dimensionality of the latent space
(ie. the result of the sample embedding).
\cr \cr
\emph{Diagnostic messages}
\cr \cr
An error is thrown if the samples stored in the \code{CellTrailsSet}
object were not embedded yet (ie. the \code{CellTrailsSet} object does not contain a
latent space matrix object; function call \code{latentSpace}
returns NULL; see \code{embedSamples}).
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)
ctset

dim(latentSpace(ctset))
\dontrun{
plot(ctset, type="latentSpace", feature_name="feature_1")
}
}
\author{
Daniel C. Ellwanger
}
