% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{removeLandmark}
\alias{removeLandmark}
\alias{removeLandmark,CellTrailsSet-method}
\title{REMOVE landmark}
\usage{
removeLandmark(object, sample_name)
}
\arguments{
\item{object}{An object of class \code{\link[CellTrails]{CellTrailsSet}}}

\item{sample_name}{A character string indicating the name of the sample}
}
\value{
An updated object of class \code{\link[CellTrails]{CellTrailsSet}}
}
\description{
Removes a landmark to a \code{CellTrailsSet} object.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Select trajectory
ctset <- selectTrajectory(ctset, component=1)

# Align samples to trajectory
ctset <- fitTrajectory(ctset)

# For illustration purposes a layout for each example dataset
# is contained in this package.
# The function trajectoryLayout allows to set a trajectory layout
# to the object; the parameter adjust='TRUE' adjusts the layout
# according to the computed pseudotime
trajectoryLayout(ctset, adjust=TRUE) <- trajLayouts$example

# Add user-defined landmark
ctset <- addLandmark(ctset, sample_name="C3_sample_10")

# Remove landmark
ctset <- removeLandmark(ctset, sample_name="C3_sample_10")
plot(ctset, type="trailblazing")
}
\author{
Daniel C. Ellwanger
}
