% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{removeTrail}
\alias{removeTrail}
\alias{removeTrail,SingleCellExperiment-method}
\title{REMOVE trail}
\usage{
removeTrail(sce, name)
}
\arguments{
\item{sce}{An object of class \code{SingleCellExperiment}}

\item{name}{Name of trail}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Removes trail from a \code{SingleCellExperiment} object.
}
\details{
\emph{Diagnostic messages}
\cr \cr
An error is thrown if the trail name is unknown. All stored trail
names can be shown using function \code{trailNames}.
}
\examples{
# Example data
data(exSCE)

# Remove trail
trailNames(exSCE)
exSCE <- removeTrail(exSCE, "Tr1")
trailNames(exSCE)
}
\seealso{
\code{trailNames} \code{addTrail}
}
\author{
Daniel C. Ellwanger
}
