% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{sampleNames,SingleCellExperiment-method}
\alias{sampleNames,SingleCellExperiment-method}
\title{GET sample names}
\usage{
\S4method{sampleNames}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A \code{character} vector
}
\description{
Retrieve sample names from a \code{SingleCellExperiment} object
}
\details{
Wrapper for \code{colnames(object)}
}
\examples{
# Generate example data
sce <- exDat()

sampleNames(sce)
}
\seealso{
\code{SingleCellExperiment}
}
\author{
Daniel C. Ellwanger
}
