% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames,SingleCellExperiment-method}
\title{GET sample names}
\usage{
sampleNames(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A \code{character} vector
}
\description{
Retrieve sample names from a \code{SingleCellExperiment} object
}
\details{
Wrapper for \code{colnames(object)}
}
\examples{
# Example data
data(exSCE)

sampleNames(exSCE)[seq_len(5)]
}
\seealso{
\code{SingleCellExperiment}
}
\author{
Daniel C. Ellwanger
}
