% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,CellTrailsSet-method}
\alias{show,CellTrailsSet-method}
\title{Shows content of a CellTrailsSet object}
\usage{
\S4method{show}{CellTrailsSet}(object)
}
\arguments{
\item{object}{A \code{CellTrailsSet} object}
}
\value{
\code{show} returns an invisible \code{NULL}
}
\description{
Shows content of a CellTrailsSet object
}
\examples{
# Generate example data
ctset <- as.CellTrailsSet(exDat())

# Show content
show(ctset)
}
\author{
Daniel C. Ellwanger
}
