% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{showTrajInfo}
\alias{showTrajInfo}
\alias{showTrajInfo,SingleCellExperiment-method}
\title{Shows relevant content of a SingleCellExperiment object for a CellTrails
analysis}
\usage{
showTrajInfo(object)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}
}
\value{
\code{showTrajInfo} returns an invisible \code{NULL}
}
\description{
Shows relevant content of a SingleCellExperiment object for a CellTrails
analysis
}
\examples{
# Example data
data(exSCE)

showTrajInfo(exSCE)
}
\author{
Daniel C. Ellwanger
}
