% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{states}
\alias{states}
\alias{states,SingleCellExperiment-method}
\title{GET states}
\usage{
states(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A \code{factor} vector
}
\description{
Retrieve computed states from a \code{SingleCellExperiment} object
}
\details{
State information is extracted from \code{colData};
factor levels are alphanumerically ordered by ID.
}
\examples{
# Generate example data
sce <- exDat()

# Assign clusters
cl <- kmeans(logcounts(sce), centers=5)$cluster
states(sce) <- cl

# Get clusters
states(sce)
}
\seealso{
\code{SingleCellExperiment} \code{findStates}
}
\author{
Daniel C. Ellwanger
}
