% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{stateTrajectoryGraph}
\alias{stateTrajectoryGraph}
\alias{stateTrajectoryGraph,CellTrailsSet-method}
\title{GET state trajectory graph}
\usage{
stateTrajectoryGraph(object)
}
\arguments{
\item{object}{An object of class \code{\link[CellTrails]{CellTrailsSet}}.}
}
\value{
An object of class \code{\link{list}}
}
\description{
Retrieve computed state trajectory graph from a \code{CellTrailsSet} object.
}
\details{
A graph is computed by spanning states maximizing the overall interface
score (i.e., number of neighboring samples between states). This method retrieves
the maximum interface spanning forest information from a \code{CellTrailsSet}
object. The return value is a \code{list} with an \code{\link{igraph}} object for each
component.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Get state trajectory graph
stateTrajectoryGraph(ctset)
}
\author{
Daniel C. Ellwanger
}
