% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{stateTrajLayout<-}
\alias{stateTrajLayout<-}
\alias{stateTrajLayout<-,SingleCellExperiment-method}
\title{SET state trajectory layout}
\usage{
stateTrajLayout(object) <- value
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}

\item{value}{A \code{ggplot} object}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Stores layout of state trajectory in \code{SingleCellExperiment} object
}
\examples{
# Example data
data(exSCE)

gp <- plotStateTrajectory(exSCE, color_by="featureName",
                          name="feature_10", recalculate=TRUE)
stateTrajLayout(exSCE) <- gp
}
\author{
Daniel C. Ellwanger
}
