% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{[,CellTrailsSet,ANY,ANY,ANY-method}
\alias{[,CellTrailsSet,ANY,ANY,ANY-method}
\title{SUBSET assay data}
\usage{
\S4method{[}{CellTrailsSet,ANY,ANY,ANY}(x, i, j, drop = "missing")
}
\arguments{
\item{x}{A \code{\link[CellTrails]{CellTrailsSet}} object}

\item{i}{Feature indices specifying elements to extract}

\item{j}{Sample indices specifying elements to extract}

\item{drop}{If TRUE the result is coerced to the lowest possible dimension (default: "missing").}
}
\value{
numerical matrix
See \code{\link[base]{drop}} for further details.
}
\description{
Extract parts of expression matrix
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Indexing
ctset[1:10, 1:10]
}
\author{
Daniel C. Ellwanger
}
