% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trailNames}
\alias{trailNames}
\alias{trailNames,SingleCellExperiment-method}
\title{GET trail names}
\usage{
trailNames(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A \code{character} vector
}
\description{
Function to extract trail names from \code{SingleCellExperiment}
object.
}
\examples{
# Generate example data
sce <- exDat()

# Reduce dimensionality
res <- embedSamples(sce)
d <- findSpectrum(res$eigenvalues, frac=30)
latentSpace(sce) <- res$components[, d]

# Find and connect states
states(sce) <- findStates(sce, max_pval=1e-3, min_feat=4)
sce <- connectStates(sce, l=15)

# Align samples to trajectory
sce <- selectTrajectory(sce, component=1)
sce <- fitTrajectory(sce)

# Add trail
sce <- addTrail(sce, "H1", "H2", "myTrail")
trailNames(sce)
}
\seealso{
\code{addTrail}
}
\author{
Daniel C. Ellwanger
}
