% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trails}
\alias{trails}
\alias{trails,SingleCellExperiment-method}
\title{GET trails}
\usage{
trails(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A DataFrame with \code{numeric} columns
}
\description{
Function to extract trail pseudotimes from a
\code{SingleCellExperiment} object.
}
\examples{
# Generate example data
sce <- exDat()

# Reduce dimensionality
res <- embedSamples(sce)
d <- findSpectrum(res$eigenvalues, frac=30)
latentSpace(sce) <- res$components[, d]

# Find and connect states
states(sce) <- findStates(sce, max_pval=1e-3, min_feat=4)
sce <- connectStates(sce, l=15)

# Align samples to trajectory
sce <- selectTrajectory(sce, component=1)
sce <- fitTrajectory(sce)

# Add trail
sce <- addTrail(sce, "H1", "H2", "myTrail")
sce <- addTrail(sce, "H1", "H3", "myTrail2")
trails(sce)
}
\seealso{
\code{addTrail}
}
\author{
Daniel C. Ellwanger
}
