% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trajComponents}
\alias{trajComponents}
\alias{trajComponents,SingleCellExperiment-method}
\title{GET trajectory component states}
\usage{
trajComponents(object)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}
}
\value{
A \code{character} vector
}
\description{
Returns states of trajectory components
\code{SingleCellExperiment} object
}
\examples{
# Example data
data(exSCE)

trajComponents(exSCE)
}
\author{
Daniel C. Ellwanger
}
