% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trajectoryFeatures<-}
\alias{trajectoryFeatures<-}
\alias{trajectoryFeatures<-,CellTrailsSet-method}
\title{SET trajectory features}
\usage{
trajectoryFeatures(object) <- value
}
\arguments{
\item{object}{An object of class \code{CellTrailsSet}}

\item{value}{A character vector containing the names of the features}
}
\value{
An updated object of class \code{CellTrailsSet}
}
\description{
Sets features that are used for trajectory reconstruction
to a \code{CellTrailsSet} object.
}
\details{
Features can be selected prior to trajectory inference.
This step is recommended as it increases the accuracy and minimzes the
runtime of the trajectory reconstruction algorithm.
\cr \cr
\emph{Diagnostic messages}
\cr \cr
The method states a warning if feature names cannot be found in the stored
assay. Since \code{CellTrailsSet} extends class \code{ExpressionSet}, all
feature names stored in a \code{CellTrailsSet} object can be retrieved by
the function \code{featureNames}. Further, this method shows a warning if
the feature selection generates samples with zero entropy, that is the
selected features were not detected in these samples.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Set trajectory features
trajectoryFeatures(ctset) <- Biobase::featureNames(ctset)[1:10]
}
\seealso{
\code{\link[CellTrails]{CellTrailsSet}}
}
\author{
Daniel C. Ellwanger
}
