% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trajectoryFeatures}
\alias{trajectoryFeatures}
\alias{trajectoryFeatures,CellTrailsSet-method}
\title{GET trajectory features}
\usage{
trajectoryFeatures(object)
}
\arguments{
\item{object}{An object of class \code{CellTrailsSet}}
}
\value{
An object of class \code{character}
}
\description{
Retrieve names of features that were selected for trajectory reconstruction
from a \code{CellTrailsSet} object.
}
\details{
Features can be selected prior to trajectory inference.
This method retrieves the user-defined features from a \code{CellTrailsSet}
object. The return value is a character vector containing the feature names.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Get trajectory features
trajectoryFeatures(ctset)
}
\author{
Daniel C. Ellwanger
}
