% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trajectoryFit}
\alias{trajectoryFit}
\alias{trajectoryFit,CellTrailsSet-method}
\title{GET trajectory fit}
\usage{
trajectoryFit(object)
}
\arguments{
\item{object}{An object of class \code{\link[CellTrails]{CellTrailsSet}}.}
}
\value{
An object of type \code{\link[base]{list}} with the following components
\describe{
  \item{\code{error}}{numerical vector; the actual distance of each sample from the fitted trajectory}
  \item{\code{traj}}{\code{\link[igraph]{igraph}} object; graph connecting samples by
  their chronological ordering}
  \item{\code{blaze}}{data.frame; indicates landmarks (branching points and end points)
  in the trajectory graph}
}
}
\description{
Retrieve trajectory fit information from a \code{CellTrailsSet} object.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Select trajectory
ctset <- selectTrajectory(ctset, component=1)

# Align samples to trajectory
ctset <- fitTrajectory(ctset)

# Get trajectory fit
trajectoryFit(ctset)
}
\author{
Daniel C. Ellwanger
}
