% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trajectoryLayout<-}
\alias{trajectoryLayout<-}
\alias{trajectoryLayout<-,CellTrailsSet-method}
\title{SET trajectory layout}
\usage{
trajectoryLayout(object, adjust) <- value
}
\arguments{
\item{object}{An object of class \code{CellTrailsSet}}

\item{adjust}{If adjustment of the layout is required. (default: TRUE)}

\item{value}{A matrix with x- and y-coordinates for each sample; rows =
samples, columns = coordinates}
}
\value{
An updated object of class \code{CellTrailsSet}
}
\description{
Sets layout which is used for trajectory visualization to a
\code{CellTrailsSet} object.
}
\details{
\emph{Diagnostic messages}
\cr \cr
An error is thrown if the number of rows of the layout does not correspond
to the number of \code{trajectorySamples} or if the number of columns does
not equal 2 or if the row names do not correspond to \code{sampleNames}.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Select trajectory
ctset <- selectTrajectory(ctset, component=1)

# Align samples to trajectory
ctset <- fitTrajectory(ctset)

# For illustration purposes a layout for each example dataset
# is contained in this package.
# The function trajectoryLayout allows to set a trajectory layout
# to the object; the parameter adjust='TRUE' adjusts the layout
# according to the computed pseudotime
trajectoryLayout(ctset, adjust=TRUE) <- trajLayouts$example

}
\seealso{
\code{\link[CellTrails]{CellTrailsSet}}
}
\author{
Daniel C. Ellwanger
}
