% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trajectoryLayout}
\alias{trajectoryLayout}
\alias{trajectoryLayout,CellTrailsSet-method}
\title{GET trajectory layout}
\usage{
trajectoryLayout(object)
}
\arguments{
\item{object}{An object of class \code{CellTrailsSet}}
}
\value{
A numeric matrix
}
\description{
Retrieve layout for trajectory visualization from
\code{CellTrailsSet} object.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Select trajectory
ctset <- selectTrajectory(ctset, component=1)

# Align samples to trajectory
ctset <- fitTrajectory(ctset)

# For illustration purposes a layout for each example dataset
# is contained in this package.
# The function trajectoryLayout allows to set a trajectory layout
# to the object; the parameter adjust='TRUE' adjusts the layout
# according to the computed pseudotime
trajectoryLayout(ctset, adjust=TRUE) <- trajLayouts$example

# Get trajectory layout
trajectoryLayout(ctset)
}
\author{
Daniel C. Ellwanger
}
