% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trajectorySamples}
\alias{trajectorySamples}
\alias{trajectorySamples,CellTrailsSet-method}
\title{GET trajectory samples}
\usage{
trajectorySamples(object)
}
\arguments{
\item{object}{An object of class \code{CellTrailsSet}}
}
\value{
An object of class \code{character}
}
\description{
Retrieve names of samples that are contained in the trajectory from
a \code{CellTrailsSet} object.
}
\details{
Samples are selected prior to trajectory inference; a state trajectory
graph may contain multiple components. Each component is composed of a set of
states. This method retrieves the names of the samples that are members of the
states contained in the user-selected state trajectory graph component from a
\code{CellTrailsSet} object.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Select trajectory
ctset <- selectTrajectory(ctset, component=1)

# Get trajectory samples
trajectorySamples(ctset)
}
\seealso{
\code{\link[CellTrails]{CellTrailsSet}}
}
\author{
Daniel C. Ellwanger
}
