% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trajectoryStates}
\alias{trajectoryStates}
\alias{trajectoryStates,CellTrailsSet-method}
\title{GET trajectory states}
\usage{
trajectoryStates(object)
}
\arguments{
\item{object}{An object of class \code{CellTrailsSet}.}
}
\value{
A \code{character} vector
}
\description{
Retrieve state IDs along a selected trajectory from a
\code{CellTrailsSet} object.
}
\details{
A state trajectory graph may contain
multiple components composed of a set of states. This method retrieves
the names of the states contained in the user-selected state trajectory
graph component from a \code{CellTrailsSet} object.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Select trajectory
ctset <- selectTrajectory(ctset, component=1)

# Get trajectory states
trajectoryStates(ctset)
}
\author{
Daniel C. Ellwanger
}
