% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trajFeatureNames<-}
\alias{trajFeatureNames<-}
\alias{trajFeatureNames<-,SingleCellExperiment-method}
\title{SET trajectory features by name}
\usage{
trajFeatureNames(object) <- value
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}

\item{value}{A character vector}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Function to set trajectory features by name
}
\examples{
# Generate example data
sce <- exDat()

# Set trajectory features
trajFeatureNames(sce) <- featureNames(sce)[1:10]
}
\author{
Daniel C. Ellwanger
}
