% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{trajLayouts}
\alias{trajLayouts}
\title{Example trajectory layouts}
\format{An object of class \code{list} of length 3.}
\usage{
data("trajLayouts")
}
\description{
This dataset contains the trajectory layouts computed for the
E15 chicken utricle and the P1 mouse utricle data sets as shown in
the original CellTrails publication.
}
\references{
Ellwanger DC, Scheibinger M, Dumont RA,
Barr-Gillespie PG, and Heller S. "Transcriptional dynamics of
hair-bundle morphogenesis revealed with CellTrails".
Cell Reports Date;Issue. doi:
}
\keyword{datasets}
