% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{userLandmarks}
\alias{userLandmarks}
\alias{userLandmarks,SingleCellExperiment-method}
\title{GET user-defined landmarks}
\usage{
userLandmarks(object)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}
}
\value{
A character vector with sample names
}
\description{
Gets user-defined landmarks from a \code{SingleCellExperiment} object.
}
\details{
Landmarks can be defined on the trajectory by the user
with \code{userLandmarks}. Landmarks can be used to extract single
trails from a trajectory.
}
\examples{
# Generate example data
sce <- exDat()

# Get landmarks
userLandmarks(sce)
}
\seealso{
\code{SingleCellExperiment}
}
\author{
Daniel C. Ellwanger
}
