% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor-methods.R
\docType{methods}
\name{as.CellTrailsSet}
\alias{as.CellTrailsSet}
\alias{as.CellTrailsSet,ExpressionSet-method}
\alias{as.CellTrailsSet,matrix-method}
\title{Coerce into CellTrailsSet}
\usage{
as.CellTrailsSet(object)

\S4method{as.CellTrailsSet}{ExpressionSet}(object)

\S4method{as.CellTrailsSet}{matrix}(object)
}
\arguments{
\item{object}{An \code{\link[Biobase]{ExpressionSet}} object or numerical matrix
with samples in rows and features in columns}
}
\value{
A \code{\link[CellTrails]{CellTrailsSet}} object
}
\description{
Function to coerce an \code{\link[Biobase]{ExpressionSet}} object or numerical
matrix into a \code{\link[CellTrails]{CellTrailsSet}} object.
}
\examples{
# Generate example data
dat_num <- simulate_exprs(10, 100)
dat_eset <- exDat()

# From numerical matrix to CellTrailsSet
ctset1 <- as.CellTrailsSet(dat_num)
ctset1

# From ExpressionSet to CellTrailsSet
ctset2 <- as.CellTrailsSet(dat_eset)
ctset2
}
\seealso{
ExpressionSet
}
