% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor-methods.R
\docType{methods}
\name{as.ExpressionSet}
\alias{as.ExpressionSet}
\alias{as.ExpressionSet,CellTrailsSet-method}
\title{Coerce CellTrailsSet into ExpressionSet}
\usage{
as.ExpressionSet(object)
}
\arguments{
\item{object}{A \code{CellTrailsSet} object}
}
\value{
An \code{ExpressionSet} object
}
\description{
Function to coerce a \code{CellTrailsSet} object into
an \code{ExpressionSet} object.
}
\details{
The resulting \code{ExpressionSet} object
contains additional phenoData information from the CellTrails
analysis (e.g., the determined states, the pseudotime
along each trail).
}
\examples{
# Generate example data
dat <- exDat()
dat
ctset <- as.CellTrailsSet(dat)

# From CellTrailsSet to ExpressionSet
dat2 <- as.ExpressionSet(ctset)
dat2
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}

ExpressionSet
}
