% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellTrailsSet}
\alias{CellTrailsSet}
\alias{CellTrailsSet-class}
\title{Class to reconstruct and visualize trajectories from biological samples}
\description{
Container for high-throughput assays and experimental metadata, as well as,
trajectory data. This class is derived from \code{\link[Biobase]{ExpressionSet}}.
}
\details{
\strong{Class-specific methods}
 \describe{
   \item{\code{show(object)}}{Informatively display object contents.}
   \item{\code{object[(index)}}{Conducts subsetting of expression data component.}
   \item{\code{trajectoryFeatures(object)}, \code{trajectoryFeatures(object)<-}}{Get/Set features for trajectory reconstruction.}
   \item{\code{latentSpace(object)}}{Get latent space from dimension dimension reduction.}
   \item{\code{eigenvalues(object)}}{Get eigenvalues from dimension reduction.}
   \item{\code{states(object)}}{Get assigned state per sample.}
   \item{\code{stateTrajectoryGraph(object)}}{Get components of the trajectory graph.}
   \item{\code{trajectoryStates(object)}}{Get states contained in selected trajectory graph component.}
   \item{\code{trajectorySamples(object)}}{Get samples contained in selected trajectory graph component.}
   \item{\code{trajectoryFit(object)}}{Get trajectory fit information.}
   \item{\code{trajectoryLayout(object)}}{Get trajectory layout information.}
 }
}
\section{Slots}{

\describe{
\item{\code{assayData}}{See \code{\link[Biobase]{ExpressionSet}}}

\item{\code{phenoData}}{See \code{\link[Biobase]{ExpressionSet}}}

\item{\code{featureData}}{See \code{\link[Biobase]{ExpressionSet}}}

\item{\code{experimentData}}{See \code{\link[Biobase]{ExpressionSet}}}

\item{\code{annotation}}{See \code{\link[Biobase]{ExpressionSet}}}

\item{\code{protocolData}}{See \code{\link[Biobase]{ExpressionSet}}}

\item{\code{useFeature}}{Object of class \code{\link[base]{logical}}, containing indices of selected trajectory features}

\item{\code{useSample}}{Object of class \code{\link[base]{logical}}, containing indices of selected trajectory samples}

\item{\code{eigen}}{Object of class \code{\link[base]{list}}, containing the latent space and eigenvalues from the spectral embedding.}

\item{\code{spanForest}}{Object of class \code{\link[base]{list}}, containing the \code{\link[igraph]{igraph}} components of the maximum interface spanning graph}

\item{\code{trajectory}}{Object of class \code{\link[base]{list}}, containing the trajectory fit}

\item{\code{layout}}{Object of class \code{\link[base]{data.frame}}, containing the geographic coordinates of the CellTrails maps}

\item{\code{trails}}{Object of class \code{\link[base]{list}}, containing trails information}
}}

\seealso{
\code{\link[Biobase]{ExpressionSet}}, \code{\link[Biobase]{eSet}}
}
\author{
Daniel C. Ellwanger
}
