% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{findSpectrum}
\alias{findSpectrum}
\alias{findSpectrum,CellTrailsSet-method}
\title{Determine number of informative latent dimensions}
\usage{
findSpectrum(ctset, frac = 100)
}
\arguments{
\item{ctset}{A \code{CellTrailsSet} object}

\item{frac}{Fraction or number (if \code{frac > 1}) of eigengaps
used to perform linear fit. (default: 100)}
}
\value{
An object of class \code{CellTrailsSpectrum}
}
\description{
Identifies the dimensionality of the latent space
}
\details{
Similar to a scree plot, this method generates a simple line
segement plot showing the lagged differences between ordered eigenvalues
(eigengaps). A linear fit is calucated on a fraction of top ranked values
to identify informative eigenvectors.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)
plot(spectr)
}
\seealso{
\code{\link[CellTrails]{CellTrailsSet}}
}
\author{
Daniel C. Ellwanger
}
