% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{fitDynamic}
\alias{fitDynamic}
\alias{fitDynamic,CellTrailsSet-method}
\title{Fit expression dynamic}
\usage{
fitDynamic(ctset, feature_name, trail_name)
}
\arguments{
\item{ctset}{A \code{CellTrailsSet} object}

\item{feature_name}{Name of feature}

\item{trail_name}{Name of trail}
}
\value{
An object of type \code{list} with the following components
\describe{
  \item{\code{pseudotime}}{The pseudotime along the trail}
  \item{\code{expression}}{The fitted expression values for each value of pseudotime}
  \item{\code{gam}}{A object of class \code{gamObject}}
}
}
\description{
Fits feature expression as a function of pseudotime along a defined trail.
}
\details{
A trail is an induced subgraph of the trajectory graph. A
trajectory graph is composed of samples (nodes) that are connected
(by weighted edges) if they are chronologically related. A trail has to be
defined by the user using \code{addTrail}. A pseudotime vector is extracted by
computing the geodesic distance for each sample from the trail's start node.
To infer the expression level of a feature as a function of
pseudotime, CellTrails used generalized additive models with a single
smoothing term with four basis dimensions. Here, for each feature CellTrails
introduces prior weights for each observation to lower the confounding
effect of drop-outs to the maximum-likelihood-based fitting process as
follows. Each non-detect of feature
\emph{j} in state \emph{h} is weighted by the relative fraction of
non-detects of feature \emph{j} in state \emph{h}; detected values are
always assigned weight = 1.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Select trajectory
ctset <- selectTrajectory(ctset, component=1)

# Align samples to trajectory
ctset <- fitTrajectory(ctset)

# For illustration purposes a layout for each example dataset
# is contained in this package.
# The function trajectoryLayout allows to set a trajectory layout
# to the object; the parameter adjust='TRUE' adjusts the layout
# according to the computed pseudotime
trajectoryLayout(ctset, adjust=TRUE) <- trajLayouts$example

# Define trail
ctset <- addTrail(ctset, from="H1", to="H3", name="Tr1")

# Fit dynamic
fit <- fitDynamic(ctset, feature_name="feature_3", trail_name="Tr1")

summary(fit)
}
\author{
Daniel C. Ellwanger
}
