% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{latentSpace}
\alias{latentSpace}
\alias{latentSpace,CellTrailsSet-method}
\title{GET latentSpace}
\usage{
latentSpace(object)
}
\arguments{
\item{object}{An object of class \code{\link[CellTrails]{CellTrailsSet}}.}
}
\value{
An object of class \code{\link[base]{matrix}}
}
\description{
Retrieve computed latent space from a \code{CellTrailsSet} object.
}
\details{
Returns the latent space computed by spectral embedding of samples. The
corresponding matrix is numeric. Rows are samples and columns are \emph{d}
components.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Get latent space
latentSpace(ctset)
}
\author{
Daniel C. Ellwanger
}
