% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{latentSpaceSNE<-}
\alias{latentSpaceSNE<-}
\alias{latentSpaceSNE<-,SingleCellExperiment-method}
\title{SET tSNE representation}
\usage{
latentSpaceSNE(object) <- value
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}

\item{value}{A \code{numeric} matrix with one column per dimension}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Stores tSNE representation in \code{SingleCellExperiment} object
}
\examples{
#' # Generate example data
sce <- exDat()

# Reduce dimensionality
res <- embedSamples(sce)
d <- findSpectrum(res$eigenvalues, frac=30)
latentSpace(sce) <- res$components[, d]

gp <- plotManifold(sce, color_by="featureName", name="feature_10")
latentSpaceSNE(sce) <- gp
}
\author{
Daniel C. Ellwanger
}
