% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{latentSpaceSNE}
\alias{latentSpaceSNE}
\alias{latentSpaceSNE,SingleCellExperiment-method}
\title{GET tSNE representation}
\usage{
latentSpaceSNE(object)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}
}
\value{
A \code{numeric} vector
}
\description{
Returns tSNE representation of latent space from
\code{SingleCellExperiment} object
}
\examples{
# Example data
data(exSCE)

latentSpaceSNE(exSCE)[seq_len(5), ]
}
\author{
Daniel C. Ellwanger
}
