% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{pca}
\alias{pca}
\alias{pca,CellTrailsSet-method}
\title{Principal Component Analysis}
\usage{
pca(ctset, do_scaling = TRUE, design = NULL)
}
\arguments{
\item{ctset}{An \code{\link{CellTrailsSet}} object}

\item{do_scaling}{FALSE = covariance matrix, TRUE = correlation matrix}

\item{design}{A numeric matrix describing the factors that should be blocked}
}
\value{
A \code{list} object containing the following components:

\item{\code{princomp}}{Principal components}

\item{\code{variance}}{Variance explained by each component}

\item{\code{loadings}}{Loading score for each feature}
}
\description{
Performs principal component analysis by spectral decomposition of
a covariance or correlation matrix
}
\details{
The calculation is done by a spectral decomposition of the
(scaled) covariance matrix of the trajectory features
as defined in the \code{CellTrails} object.
Features with zero variance get automatically removed. To account for systematic bias in the
expression data (e.g., cell cycle effects), a design matrix can be
provided for the learning process. It should list the factors that should be blocked and
their values per sample. It is suggested to construct a design matrix with
\code{\link[stats]{model.matrix}}.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Principal component analysis
pca_result <- pca(ctset)

barplot(pca_result$variance[1:10], ylab="Variance",
        names.arg=colnames(pca_result$princomp)[1:10], las=2)
plot(pca_result$princomp, xlab="PC1", ylab="PC2")
}
\author{
Daniel C. Ellwanger
}
