% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plotStateSize}
\alias{plotStateSize}
\alias{plotStateSize,SingleCellExperiment-method}
\title{Visualize the number of samples per state}
\usage{
plotStateSize(sce)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}
}
\value{
A \code{ggplot} object
}
\description{
Shows barplot of state size distribution
}
\details{
Barplot showing the absolute number of samples per state.
}
\examples{
# Generate example data
sce <- exDat()

# Define states
X <- logcounts(sce)
X <- t(X)
states(sce) <- kmeans(X, centers=5)$cluster
plotStateSize(sce)
}
\seealso{
\code{ggplot} \code{states}
}
\author{
Daniel C. Ellwanger
}
