% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plotTrajectoryFit}
\alias{plotTrajectoryFit}
\alias{plotTrajectoryFit,SingleCellExperiment-method}
\title{Visualize trajectory fit residuals}
\usage{
plotTrajectoryFit(sce)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}
}
\value{
A \code{ggplot} object
}
\description{
Method visualizes the fitting residuals along the
trajectory backbone.
}
\details{
Shows a single tree component of the computed trajectory graph.
Each point in this plot represents a state and can be colorized
according to feature expression (mean expression per state) or experimental
metadata (arithmetic mean or percentage distribution of categorial values).
The component is defined by parameter \code{component}. If the trajectory
graph contains only a single component, then this parameter can be left
undefined. The points' coloration can be defined via the attributes
\code{feature_name} or \code{pheno_type}. Missing sample lables are
recovered using nearest neighbor learning.
}
\examples{
# Example data
data(exSCE)

plotTrajectoryFit(exSCE)
}
\seealso{
\code{fitTrajectory} \code{trajResiduals}
}
\author{
Daniel C. Ellwanger
}
