% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{removeTrail}
\alias{removeTrail}
\alias{removeTrail,SingleCellExperiment-method}
\title{REMOVE trail}
\usage{
removeTrail(sce, name)
}
\arguments{
\item{sce}{An object of class \code{SingleCellExperiment}}

\item{name}{Name of trail}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Removes trail from a \code{SingleCellExperiment} object.
}
\details{
\emph{Diagnostic messages}
\cr \cr
An error is thrown if the trail name is unknown. All stored trail
names can be shown using function \code{trailNames}.
}
\examples{
# Generate example data
sce <- exDat()

# Reduce dimensionality
res <- embedSamples(sce)
d <- findSpectrum(res$eigenvalues, frac=30)
latentSpace(sce) <- res$components[, d]

# Find and connect states
states(sce) <- findStates(sce, max_pval=1e-3, min_feat=4)
sce <- connectStates(sce, l=15)

# Align samples to trajectory
sce <- selectTrajectory(sce, component=1)
sce <- fitTrajectory(sce)

# Add trail
sce <- addTrail(sce, "H1", "H2", "myTrail")
trailNames(sce)

# Remove trail
sce <- removeTrail(sce, "myTrail")
trailNames(sce)

}
\seealso{
\code{trailNames} \code{addTrail}
}
\author{
Daniel C. Ellwanger
}
