% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{removeTrail}
\alias{removeTrail}
\alias{removeTrail,CellTrailsSet-method}
\title{REMOVE trail}
\usage{
removeTrail(ctset, name)
}
\arguments{
\item{ctset}{An object of class \code{CellTrailsSet}}

\item{name}{Name of trail}
}
\value{
An updated object of class \code{CellTrailsSet}
}
\description{
Removes trail from a \code{CellTrailsSet} object.
}
\details{
\emph{Diagnostic messages}
\cr \cr
An error is thrown if the trail name is unknown. All stored trail names can be shown
using function \code{trailNames}.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Select trajectory
ctset <- selectTrajectory(ctset, component=1)

# Align samples to trajectory
ctset <- fitTrajectory(ctset)

# For illustration purposes a layout for each example dataset
# is contained in this package.
# The function trajectoryLayout allows to set a trajectory layout
# to the object; the parameter adjust='TRUE' adjusts the layout
# according to the computed pseudotime
trajectoryLayout(ctset, adjust=TRUE) <- trajLayouts$example

# Define trail
ctset <- addTrail(ctset, from="H1", to="H3", name="Tr1") # Define trail
ctset <- addTrail(ctset, from="H1", to="H4", name="Tr2") # Define trail
trailNames(ctset)

# Remove trail
ctset <- removeTrail(ctset, name="Tr1")
trailNames(ctset)
}
\author{
Daniel C. Ellwanger
}
