% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{selectTrajectory}
\alias{selectTrajectory}
\alias{selectTrajectory,CellTrailsSet-method}
\title{Select component from trajectory graph}
\usage{
selectTrajectory(ctset, component)
}
\arguments{
\item{ctset}{A \code{CellTrailsSet} object}

\item{component}{Number of component to be selected}
}
\value{
An updated \code{CellTrailsSet} object
}
\description{
Retains a single component of a trajectory graph.
}
\details{
The construction of a trajectory graph may result in a forest
having multiple tree components, which may represent individual
trajectories or isolated nodes. This method should be used to extract a
single component from the graph. A component is
identified by its (integer) number.
\cr \cr
\emph{Diagnostic messages}
\cr \cr
An error is thrown if the states have not been connected yet;
function \code{connectStates}
needs to be called first. An error is thrown if an unknown
component (number) is selected.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Select trajectory
ctset <- selectTrajectory(ctset, component=1)

stateTrajectoryGraph(ctset)
trajectorySamples(ctset)
}
\author{
Daniel C. Ellwanger
}
