% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{states}
\alias{states}
\alias{states,SingleCellExperiment-method}
\title{GET states}
\usage{
states(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A \code{factor} vector
}
\description{
Retrieve computed states from a \code{SingleCellExperiment} object
}
\details{
State information is extracted from \code{colData};
factor levels are alphanumerically ordered by ID.
}
\examples{
# Example data
data(exSCE)

states(exSCE)[seq_len(5)]
}
\seealso{
\code{SingleCellExperiment} \code{findStates}
}
\author{
Daniel C. Ellwanger
}
