% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{states}
\alias{states}
\alias{states,CellTrailsSet-method}
\title{GET states}
\usage{
states(object)
}
\arguments{
\item{object}{An object of class \code{CellTrailsSet}.}
}
\value{
A \code{factor} vector
}
\description{
Retrieve computed states from a \code{CellTrailsSet} object.
}
\details{
State information is extracted from \code{featureData};
factor levels are alphanumerically ordered by ID.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Get state assignments
states(ctset)
}
\author{
Daniel C. Ellwanger
}
