% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trailNames<-}
\alias{trailNames<-}
\alias{trailNames<-,SingleCellExperiment-method}
\title{SET trail names}
\usage{
trailNames(object) <- value
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}

\item{value}{A character vector with the trail names}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Enables to rename trails stored in a \code{SingleCellExperiment}
object.
}
\details{
\emph{Diagnostic messages}
\cr \cr
An error is thrown if the number of names does not correspond to the number
of trails stored in the object. Further, trail names are required
to be unique.
}
\examples{
# Generate example data
sce <- exDat()

# Reduce dimensionality
res <- embedSamples(sce)
d <- findSpectrum(res$eigenvalues, frac=30)
latentSpace(sce) <- res$components[, d]

# Find and connect states
states(sce) <- findStates(sce, max_pval=1e-3, min_feat=4)
sce <- connectStates(sce, l=15)

# Align samples to trajectory
sce <- selectTrajectory(sce, component=1)
sce <- fitTrajectory(sce)

# Add trail
sce <- addTrail(sce, "H1", "H2", "myTrail")
trailNames(sce)
trailNames(sce) <- "ABC"
trailNames
}
\seealso{
\code{addTrail}
}
\author{
Daniel C. Ellwanger
}
