% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trailNames}
\alias{trailNames}
\alias{trailNames,CellTrailsSet-method}
\title{GET trail names}
\usage{
trailNames(object)
}
\arguments{
\item{object}{An object of class \code{CellTrailsSet}}
}
\value{
A \code{character} vector
}
\description{
Get names of trails from \code{CellTrailsSet} object.
}
\examples{
# Generate example data
dat <- exDat()

# Create container
ctset <- as.CellTrailsSet(dat)

# Embed samples
ctset <- embedSamples(ctset)

# Find spectrum
spectr <- findSpectrum(ctset, frac=25)

# Reduce dimensionality
ctset <- reduceDimensions(ctset, spectr)

# Find states
ctset <- findStates(ctset, max_pval=1e-3, min_feat=4)

# Connect states
ctset <- connectStates(ctset, l=20)

# Select trajectory
ctset <- selectTrajectory(ctset, component=1)

# Align samples to trajectory
ctset <- fitTrajectory(ctset)

# For illustration purposes a layout for each example dataset
# is contained in this package.
# The function trajectoryLayout allows to set a trajectory layout
# to the object; the parameter adjust='TRUE' adjusts the layout
# according to the computed pseudotime
trajectoryLayout(ctset, adjust=TRUE) <- trajLayouts$example

# Define trail
ctset <- addTrail(ctset, from="H1", to="H3", name="Tr1") # Define trail
ctset <- addTrail(ctset, from="H1", to="H4", name="Tr2") # Define trail

# Get trail names
trailNames(ctset)
}
\author{
Daniel C. Ellwanger
}
