% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trailNames}
\alias{trailNames}
\alias{trailNames,SingleCellExperiment-method}
\title{GET trail names}
\usage{
trailNames(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A \code{character} vector
}
\description{
Function to extract trail names from \code{SingleCellExperiment}
object.
}
\examples{
# Example data
data(exSCE)

trailNames(exSCE)
}
\seealso{
\code{addTrail}
}
\author{
Daniel C. Ellwanger
}
