% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trails}
\alias{trails}
\alias{trails,SingleCellExperiment-method}
\title{GET trails}
\usage{
trails(object)
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}
}
\value{
A DataFrame with \code{numeric} columns
}
\description{
Function to extract trail pseudotimes from a
\code{SingleCellExperiment} object.
}
\examples{
# Example data
data(exSCE)

trails(exSCE)
}
\seealso{
\code{addTrail}
}
\author{
Daniel C. Ellwanger
}
