% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trajLayout<-}
\alias{trajLayout<-}
\alias{trajLayout<-,SingleCellExperiment-method}
\title{SET trajectory layout}
\usage{
trajLayout(object, adjust) <- value
}
\arguments{
\item{object}{An object of class \code{SingleCellExperiment}}

\item{adjust}{Indicates if layout has to be adjusted such that edge lengths
correlate to pseudotime (default: TRUE)}

\item{value}{A data.frame with x- and y-coordinates for
each sample (rows = samples, columns = coordinates)}
}
\value{
An updated object of class \code{SingleCellExperiment}
}
\description{
Sets layout used for trajectory visualization to a
\code{SingleCellExperiment} object.
}
\details{
CellTrails implements a module which can incorporate pseudotime information
into the the graph layout (activated via parameter \code{adjust}). Here,
edge lengths between two nodes (samples)
will then correspond to the inferred pseudotime that separates two samples
along the trajectory.
\cr \cr
\emph{Diagnostic messages}
\cr \cr
An error is thrown if the number of rows of the layout does not correspond
to the number of trajectory samples or if the number of columns is
less than 2, or if the row names do not correspond to \code{sampleNames}.
}
\examples{
# Create example data
sce <- exDat()

# Sample embedding and trajectory fit
res <- embedSamples(sce)
d <- findSpectrum(res$eigenvalues, frac=30)
latentSpace(sce) <- res$components[, d]
states(sce) <- findStates(sce, max_pval=1e-3, min_feat=4)
sce <- connectStates(sce, l = 15)
sce <- selectTrajectory(sce, component=1)
sce <- fitTrajectory(sce)

# For illustration purposes a graphml file for the example dataset
# is contained in this package.
# The function trajLayout allows to set a trajectory layout
# to the object; the parameter adjust='TRUE' adjusts the layout
# to represent the computed pseudotime
fn <- system.file("exdata", "exdat.graphml", package="CellTrails")
tl <- read.ygraphml(fn)
trajLayout(sce, adjust=TRUE) <- tl
}
\seealso{
\code{write.ygraphml} \code{trajSampleNames}
}
\author{
Daniel C. Ellwanger
}
