% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{userLandmarks<-}
\alias{userLandmarks<-}
\alias{userLandmarks<-,SingleCellExperiment-method}
\title{SET user-defined landmarks}
\usage{
userLandmarks(object) <- value
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}

\item{value}{A character vector with sample names}
}
\value{
An updated \code{SingleCellExperiment} object
}
\description{
Set user-defined landmarks to a \code{SingleCellExperiment} object.
}
\details{
Landmarks can be defined on the trajectory and can be necessary to
extract individual trails from a trajectory.
\cr \cr
\emph{Diagnostic messages}
\cr \cr
An error is thrown if the trajectory has not been reconstructed yet.
}
\examples{
# Generate example data
sce <- exDat()

# Reduce dimensionality
res <- embedSamples(sce)
d <- findSpectrum(res$eigenvalues, frac=30)
latentSpace(sce) <- res$components[, d]

# Find and connect states
states(sce) <- findStates(sce, max_pval=1e-3, min_feat=4)
sce <- connectStates(sce, l=15)

# Align samples to trajectory
sce <- selectTrajectory(sce, component=1)
sce <- fitTrajectory(sce)

# Set landmarks
userLandmarks(sce) <- sampleNames(sce)[5:7]
}
\seealso{
\code{SingleCellExperiment} \code{fitTrajectory}
}
\author{
Daniel C. Ellwanger
}
